requirements.txt
ftfy==6.3.1
matplotlib==3.7.5
numpy==1.24.3
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==11.1.0
regex==2024.9.11
scikit_learn==1.3.2
scipy==1.15.1
setuptools==75.1.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.66.5
transformers==4.45.1
